SW PPL is working
