+++ examples/asrc_demo/bsp_config/XK_VOICE_L71/platform/platform_init.c
@@ -114,6 +114,7 @@ static void usb_init(void)
 void platform_init(chanend_t other_tile_c)
 {
     rtos_intertile_init(intertile_ctx, other_tile_c);
+
     rtos_intertile_init(intertile_usb_audio_ctx, other_tile_c);
     rtos_intertile_init(intertile_i2s_audio_ctx, other_tile_c);
