+++ examples/ffva/bsp_config/XK_VOICE_L71/platform/driver_instances.c
@@ -38,3 +38,6 @@ rtos_dfu_image_t *dfu_image_ctx = &dfu_image_ctx_s;
 
 static rtos_uart_tx_t uart_tx_ctx_s;
 rtos_uart_tx_t *uart_tx_ctx = &uart_tx_ctx_s;
+
+static sw_pll_ctx_t sw_pll_ctx_s;
+sw_pll_ctx_t *sw_pll_ctx = &sw_pll_ctx_s;
